added exception

+++ book-api/app/Controllers/Http/BooksController.ts
@@ -3,6 +3,7 @@ import CreateBookValidator from 'App/Validators/CreateBookValidator'
 import Book
  from 'App/Models/Book'
 import UpdateBookValidator from 'App/Validators/UpdateBookValidator'
+import UnauthorizedException from 'App/Exceptions/UnauthorizedException'
 export default class BooksController {
     public async index() {
         // const books = await Book.all()
@@ -17,18 +18,16 @@ export default class BooksController {
         const validatedData = await request.validate(CreateBookValidator)
 
         // user type must be Admin to create book
-        const userType = auth.user?.userType
-
-        if (userType === 'Admin') {
+        if (auth.user?.userType !== 'Admin') {
+            throw new UnauthorizedException('This user is not authorized to create a book record.')
+        } else {
             const book = await auth.user?.related('books').create(validatedData)
             console.log(auth.user?.toJSON())    // to extract user info
             // fetch user that created the book and category it belongs to
             await book?.load('user')
             await book?.load('category')
             return book
-        } else {
-            return 'This user is not authorized to create a book record.'
-        }
+        } 
     }
 
     public async show ({ params }:HttpContextContract) {
@@ -41,24 +40,30 @@ export default class BooksController {
         return book
     }
 
-    public async update ({ request, params }:HttpContextContract) {
+    public async update ({ request, params, auth }:HttpContextContract) {
         const book = await Book.findOrFail(params.id)   // fetch book to update
-
         const validatedData = await request.validate(UpdateBookValidator)
 
+        if (auth.user?.userType !== 'Admin') {
+            throw new UnauthorizedException('This user is not authorized to update a book record.')
+        }
         // update book
         book.merge(validatedData)
         await book.save()   // pass changes to db
 
         // fetch user that created the book and category it belongs to
-        await book.preload('user')
-        await book.preload('category')
+        await book.load('user')
+        await book.load('category')
 
         return book
     }
 
-    public async destroy({ params }:HttpContextContract) {
+    public async destroy({ params, auth }:HttpContextContract) {
         const book = await Book.findOrFail(params.id)      // fetch thread to delete
+
+        if (auth.user?.userType !== 'Admin') {
+            throw new UnauthorizedException('This user is not authorized to delete a book record.')
+        }
 
         return book.delete()
     }
